added the sweet alerts

+++ src/layouts/MyCart.jsx
@@ -3,6 +3,7 @@ import { useState } from 'react';
 
 import { useLoaderData } from 
 'react-router-dom';
+import Swal from 'sweetalert2';
 
 const MyCart = () => {
     const cart=useLoaderData();
@@ -12,24 +13,60 @@ const MyCart = () => {
     console.log(carts);
     const handleDelete=(_id)=>{
         console.log("delete",_id);
-        fetch(`http://localhost:5000/cart/${_id}`,
-        {
-            method:'DELETE'
-        })
-        .then(res=>res.json())
-        .then(result=>{
-            console.log(result);
-            if(result.deletedCount>0)
+        Swal.fire({
+          title: 'Are you sure?',
+          text: "You won't be able to revert this!",
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonColor: '#3085d6',
+          cancelButtonColor: '#d33',
+          confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+          if (result.isConfirmed) {
+            Swal.fire(
+              'Deleted!',
+              'Your file has been deleted.',
+              'success'
+            )
+            fetch(`https://media-store-server-ffaumifm6-nismahossain41982-gmailcom.vercel.app/cart/${_id}`,
             {
-                alert("Deleted from Cart");
-
-                const remainingCarts=
-            carts.filter(cart=>cart._id!==_id);
-
-                setCarts(remainingCarts);
-            }
-           
+                method:'DELETE'
+            })
+            .then(res=>res.json())
+            .then(result=>{
+                console.log(result);
+                if(result.deletedCount>0)
+                {
+                    
+    
+                    const remainingCarts=
+                carts.filter(cart=>cart._id!==_id);
+    
+                    setCarts(remainingCarts);
+                }
+               
+            })
+          }
         })
+        
+        // fetch(`https://media-store-server-ffaumifm6-nismahossain41982-gmailcom.vercel.app/cart/${_id}`,
+        // {
+        //     method:'DELETE'
+        // })
+        // .then(res=>res.json())
+        // .then(result=>{
+        //     console.log(result);
+        //     if(result.deletedCount>0)
+        //     {
+        //         alert("Deleted from Cart");
+
+        //         const remainingCarts=
+        //     carts.filter(cart=>cart._id!==_id);
+
+        //         setCarts(remainingCarts);
+        //     }
+           
+        // })
     }
